refactor(agent): extract createAgent helpers into buildAgentDeps
Reduces `createAgent`'s cyclomatic complexity from 59 to well under 15
by lifting the resolver chain + the conditional-spread `Agent`
options assembly into `src/agent/internal/buildAgentDeps.ts`.

`createAgent` is now a flat orchestrator:

  1. `buildAgentDeps(config)` — single-pass resolution of every
     subsystem, returning a `ResolvedDeps` bag.
  2. `installModuleSkills(skills, modules)` — module-contributed
     skill installation with consumer-id precedence preserved.
  3. `new Agent(toAgentDependencies(deps, config))` — assembly of
     the `AgentDependencies` shape, with the conditional spreads
     chunked into per-subsystem helpers (`portOptions`,
     `lifecycleOptions`, `worldOptions`, `persistenceOptions`,
     `cognitionOptions`) so each chunk's complexity stays small.
  4. `applyPassiveModifiers(agent, sd)` — species-declared passive
     modifier dispatch on the constructed agent.

No public API changes; the `CreateAgentConfig` shape is unchanged
and the helper file lives under `internal/` (per CLAUDE.md, not
re-exported from the barrel).

Roadmap row 7 of `docs/plans/2026-04-25-comprehensive-polish-and-harden.md`.
